Moved new auto-diff backend to src
